refactor: formalize device backend boundary
