+++ custom_components/battery_smartflow_ai/core/ports/__init__.py
@@ -1,6 +1,7 @@
 """Platform-independent side-effect boundaries for the BSFAI core."""
 
 from .clock import Clock
+from .device_backend import DeviceBackend, DeviceBackendExecutionError
 from .state_store import (
     StateLoadResult,
     StateSaveResult,
@@ -10,6 +11,8 @@ from .state_store import (
 
 __all__ = [
     "Clock",
+    "DeviceBackend",
+    "DeviceBackendExecutionError",
     "StateLoadResult",
     "StateSaveResult",
     "StateStore",
